+++ internal/usecase/movement_usecase.go
@@ -2,6 +2,7 @@ package usecase
 
 import (
 	"context"
+	"time"
 
 	"personal-finance/internal/domain"
 	"personal-finance/internal/infrastructure/repository/transaction"
@@ -13,10 +14,12 @@ import (
 type (
 	MovementRepository interface {
 		Add(ctx context.Context, tx *gorm.DB, movement domain.Movement) (domain.Movement, error)
+		FindByPeriod(ctx context.Context, period domain.Period) (domain.MovementList, error)
 	}
 
 	RecurrentRepository interface {
 		Add(ctx context.Context, tx *gorm.DB, recurrent domain.RecurrentMovement) (domain.RecurrentMovement, error)
+		FindByMonth(ctx context.Context, month time.Time) ([]domain.RecurrentMovement, error)
 	}
 
 	Movement struct {
@@ -117,4 +120,42 @@ func (u *Movement) Add(ctx context.Context, movement domain.Movement) (domain.Mo
 	}
 
 	return result, nil
+}
+
+func (u *Movement) FindByPeriod(ctx context.Context, period domain.Period) ([]domain.Movement, error) {
+	movements, err := u.movementRepo.FindByPeriod(ctx, period)
+	if err != nil {
+		return []domain.Movement{}, domain.WrapInternalError(err, "error to find transactions")
+	}
+
+	recurrents, err := u.recurrentRepo.FindByMonth(ctx, period.To)
+	if err != nil {
+		return nil, domain.WrapInternalError(err, "error to find recurrents")
+	}
+
+	return u.mergeMovementsWithRecurrents(movements, recurrents, period.To), nil
+}
+
+func (u *Movement) mergeMovementsWithRecurrents(
+	movements domain.MovementList,
+	recurrents []domain.RecurrentMovement,
+	date time.Time,
+) []domain.Movement {
+	recurrentMap := make(map[uuid.UUID]struct{}, len(recurrents))
+	for i, mov := range movements {
+		if mov.RecurrentID != nil {
+			movements[i].IsRecurrent = true
+			recurrentMap[*mov.RecurrentID] = struct{}{}
+		}
+	}
+
+	for _, recurrent := range recurrents {
+		if _, ok := recurrentMap[*recurrent.ID]; !ok {
+			mov := domain.FromRecurrentMovement(recurrent, date)
+			mov.ID = mov.RecurrentID
+			movements = append(movements, mov)
+		}
+	}
+
+	return movements
 }